Refactored FrameBufferObject so that the Lwjgl-specific implementation is separated from a generic interface that is used by CanvasRenderer.
